# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для микро сервиса",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Основной API"
        },
        {
            "url": "http://test-api.php-cat.com",
            "description": "Боевой домен API"
        }
    ],
    "paths": {
        "/api/notes": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "получаем список записей",
                "responses": {
                    "200": {
                        "description": "норм ответ"
                    },
                    "401": {
                        "description": "Ошибочка"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "создание записки",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "название (латиница, цифры)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "описание",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "добавили, окей"
                    },
                    "401": {
                        "description": "Ошибочка"
                    }
                }
            }
        },
        "/api/notes/{id}/edit": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "создание записки",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "номер записи (id)",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "название (латиница, цифры)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "описание",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "добавили, окей"
                    },
                    "401": {
                        "description": "Ошибочка"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Notes",
            "description": "Работа с записками"
        }
    ],
    "security": [
        []
    ]
}